Draw new mission markers for all new missions, not just the current selection.
This lets you see all missions and their locations at once. We have
one significant problem that needs to be fixed: this makes it almost
impossible to discern where a mission you selected is, if there's a
particular mission you think you want to look at. Will need a custom
solution for this.