Completely gutted the Lua outfits system.
We have no plans for using the Lua outfit system, and it exists in
Naikari in a highly unfinished state (from long before bobbens
actually started adding Lua outfits to Naev). While the idea isn't
altogether terrible, in practice it's a major mess and makes outfits
needlessly complicated, especially considering the kind of game we
want Naikari to be. While we may want to add e.g. more complex
weapon mechanics in the future, these would be better implemented
as standard mechanics in C, rather than as special mechanics that
force outfits to explain in "extra description" text how they work.
(Not to mention, Naev has gotten a lot slower in the past few years,
and I'm sure Lua outfits haven't helped with that.)