Changed the C-side sorting for pilot.actives() to make more sense.
GUI code is actively choosing to ignore sorting that it's directly
asking for solely on the basis that its sorting didn't make any sense;
it was sorting the outfits by weapset, which makes sense, but higher
weapset numbers were placed first.

This fixes that. Now, no weapset is treated specially so outfits with
no weapset are still placed last, but for the rest, lower weapset
outfits are placed in front.