Neo: Added a function for rendering a targeted pilot (+ some other stuff).
The main objective of this commit was to add a function for rendering
a targeted pilot's "icon" (ship sprite). Along the way, I realized that
I needed to expose a line-drawing API to Lua (since it wasn't exposed
before). I also noticed a mistake in the Brushed GUI's programming
which never caused a problem only because a certain condition always
turned out to be true, and fixed it.

I haven't tested the new render_pilotIcon function in the Neo GUI,
but there's at least no syntax errors. (Will actually test it when I
get to constructing the target indicator.)