Replaced special handling of radar right-click with right-click follow option.
The system where right-clicking a target follows them only if they're
selected is rather clunky. I recall doing that for a good reason, but
nowadays it's probably best to have a simpler "right-click follow"
option instead. Leaving it off by default for now; will experiment
with how it feels in the lead-up to the next release.