Started making changes away from system claims and toward pilot-centric methods.
This is going to take awhile, but I want to make it so that the primary
means of avoiding conflict between missions and events happens at the
pilot level, not at the system level. This is the plan:

1. We'll use the pilot variable "natural" to determine whether a
   pilot is considered to be "free" for use by missions and events.
   This is something bobbens added a long time ago, but that we
   never used. Have made it so that mercenaries are "natural" too,
   and so that the ship stealing mission checks if a pilot is "natural"
   rather than the custom "nosteal".
2. We'll make judicious use of pilot.setNoClear() for mission pilots
   that should not removed by other missions in the same system.

Once we're done with all this, it should be possible to relegate
system claims to very special cases.