Removed outdated "target GFX" APIs which did nothing useful.
I don't think this has ever been true as long as I've known about Naev,
but I guess some time in the distant past the target indicators were
textures rather than the shaders that exist now. For some reason, thô,
when that was changed, whoever made the change decided to just leave
this obsolete "target GFX" API in and not only that, but leave it as
something that's mandatory to display target graphics, despite the
fact that these textures that get loaded are literally unused. The
comments said something about "the player is probably dead", but
the only thing that might have been "solved" is the target indicator
still displaying while the player is in the process of exploding.
That's not a huge deal, and even if it was, holding onto an obsolete
API that pulls unused textures isn't the way to solve it.