Changed how ai.land() is handled and added ai.canLand().
The previous implementation errored when there was no land target,
but permitted other unlandable conditions to pass with a simple
failure return value. There was also no code in place to allow
checking for whether landing in the current state was possible;
the codebase just depended on the assumption that the previous
planet selection is still valid. It usually was, but I ran into one
case where an AI got stuck at this for one reason or another and
then started printing repeated warnings because, lacking a fail
condition, it just kept trying over and over again.

So now, ai.land() always errors when landing is impossible, and a
new function, ai.canLand(), has been added to check for that state
without causing errors. Currently, when it does encounter that, it
will print a warning and then pop the task. This doesn't exactly
solve the case I found (where the code messed up and led to the
__landgo subtask being pushed when there was no valid land target),
but it at least fails more gracefully and will give a more useful
warning next time something like this happens.