Sort the faction stack to improve efficiency.
This was a bit of an undertaking, but should definitely be worth it.
Profiling showed that while it doesn't use much CPU time per call,
faction_getStackPos() is one of the most used functions in the entire
codebase, accounting for 9% of all time spent. One major inefficiency
with it was that despite the fact that you would only encounter 2–4
factions in a typical system at once, and despite that, you might
have to go thrû nearly the entire faction list depending on where
they happened to be in the stack.

This change simply sorts the faction stack when entering a new system
based on the factions' presence in the system: factions with higher
presence in the system are closer to the top of the faction stack,
meaning their stack position can be identified quicker (sometimes
much quicker).