Made collision simplification more aggressive and potentially more efficient.
It now goes to simple collision mode (where only the weapon's specific
target is checked) when its distance from the player is more than
the screen width or height (whichever is bigger). It also checks this
before checking if the weapon is one of the player's weapons, since
the number of AI weapons is almost always greater than the number of
player weapons.