Changed the methods used by brakeCheckReverseThrusters and minbrakedist.
The formula bobbens implemented seems to be based directly off of
what was used for the "inaccurate" version of ai.minbrakedist, which
seems to be worse than the method that's been in use for autonav for
years now. I don't understand the math behind it all, but I saw
a clearly much stronger tendency to overshoot jump points when
autonav was tuned to use that algorithm.