Properly update things when changing scale factor.
There was a little message that told you you needed to restart the
game for the scale factor to take effect. This wasn't actually true;
the scale factor would update the next time the game knew to update
for the window size. Got rid of that message and added and used a
pathway allowing things to get updated properly even if nothing but
scale factor was changed.