"thrû" → "thru"
Even thô it's a spelling I came up with, I'm chilling on "thrû". "û"
just looks real weird visually and is hard to process, I guess because
the breve accent looks too similar to a tilde in that context, maybe
in combination with a visual similarith between "u" and "n" (which
might be making me fight my brain to not read it as something like ñ).