Mission computer: added a check I forgot and simulate click when accept.
I forgot to add a check for whether or not a map click regenerated
the list, so I added that. I also made it do the selection of a
hilighted mission when the list is first generated and made the
mission accept aftercare for the mission computer also simulate a
map click.