Simplified and improved the code that decided whether to do simple collisions.
I swear we weren't high when we made it check distance from the player
rather than just converting to screen coördinates. 😆 Anyway, it just
converts the weapon's position to screen coördinates and compares that
to the screen bounds now. Also changed it to be a little more aggressive,
simplifying absolutely everything outside of view (and not adding an
extra buffer).

Other than making things a tad more efficient, this more importantly
avoids the possibility of the collision simplification becoming
noticeable while the camera is script-controlled.