Reorganized AI hyperspace code a bit to make it more robust.
An unfortunate side-effect of this change is that AI pilots will
now always brake, even if they have instant jump. I tried for like
an hour to think of a way to avoid this, but I just couldn't, since
there's no obvious way to tell in that case whether it failed or
just needs more time.