Added internal support for hilighting planets and jump points.
So turns out it was actually the "hilight" shader Naev was using,
not the "hilight_circle" shader. Oh well. Imported that as well.

This implementation of "hilighting" planets and jump points is a
little bit different from Naev's implementation: instead of extending
mission markers, it actually follows a similar convention to pilot
hilights. This has a weakness in that it could cause one mission to
remove another mission's planet hilight. Will need to decide on what
to do to avoid this problem; we might want to do roughly the same
thing Naev does, just with a different API and supporting events as
well as missions.