Removed "hypergates" and everything to do with them.
Our "hypergates" were based on QuasiSpace from Star Control II. In
that game, QuasiSpace is a useful idea that cuts travel time and,
even more importantly, fuel use substantially. For Naikari, thô,
what we're finding is it just doesn't do anything meaningful.

There are a couple of things working against this system in Naikari.
For starters, since it only allows you to go to places you've already
been, it doesn't have a role in exploration, just standard travel.
This is partly because we want exploration to be intimate (not just
"teleport to new place", but rather "discover new place naturally"),
but it's also necessitated for story reasons if the Empire is
responsible for constructing them mid-game.

Another factor that makes this system less good than QuasiSpace is
that you can't instantly teleport to the hub zone, which roughly
halves the travel time benefit. This doesn't have to be a huge
problem in theory; LOTRO, for example, has hubs you can go to in
order to instant-travel to other hubs, and this is common in MMORPGs.

However, the fatal flaw with the system in Naikari is simply that
normal travel times are so short that the reduction afforded by a
quick travel hub system just isn't worth much. I decided to really
test this out by flying from the south extreme edge of the map to the
top extreme edge with and without the assistence of hypergates and
hidden jump points. Unassisted, this worst-case-scenario travel time
was five minutes. Assisted, it was two minutes. That's a very
significant relative reduction, but a rather minor absolute reduction.

And in most cases, travel needs are far less extreme than that. What
we've found in playtesting is that even when we're outright given
knowledge of hypergates, we *never* have a reason to use them. The
simple fact is that most of the time, the place you need to go to
isn't all the way across the map; it's just a few jump points away.
Even a Schroedinger feels like overkill when you consider this, let
alone a special fast-travel hub system.

Plus, story-wise, "hypergates" feel like a cop-out. Realistically,
the Empire has no legitimate motivation to build them, and the rest
of the factions have no motivation either. The only reason for
"hypergates" to exist is to allow the Proteron to come along and
start causing problems everywhere in the galaxy at once, which is
a bit overly convenient. I think it would be much more interesting
for there to be some "geographical" way the Proteron are able to
invade.

Of course, this also means we'll have to come up with a different
reason for the Incident. I always felt the "hypergates exploded"
explanation felt hollow anyhow, thô.