Removed the superfluous TARGET text that gets in the way.
I tried to also fix up the cross rendering so that it shows a proper
outline, but there's something wrong with the shader that I can't
exactly diagnose. I did do the thing that's supposed to cause the
cross to have an outline, but it doesn't entirely work for some
reason.