Started to reverse-engineer these cryptic shaders and make them legible.
bobbens, writing shaders, has a very bad habit of using cryptic
variable names and not commenting his code at all, which has always
turned us off from actually working on shader code. But I know for
a fact that shader code can be written in a way that's understandable.
This is the start of an effort to rewrite all shader code with that
goal in mind. I started with asteroidmarker.frag because it's a
relatively simple shader that happens to be toward the start of the
alphabet. Was able to reverse-engineer most of it, but still haven't
gotten a complete understanding of everything (in particular, I
still don't understand the internals of sdBox() or the meaning behind
the variable "uv"). This is something I'll work on gradually.

A major goal of this would be to fix a problem with the crosshair
shader which leads to a very thin outline on my system. I'd also
like to get rid of cases where the shader code intentionally does
something with undefined behavior (and which includes division by
zero), since that practice could have compatibility issues.