Finished reverse-engineering and documenting asteroidmarker.
Looks like I got something wrong before, but that's ok. I finished
reverse-engineering it and sdBox(), and I modified it so that it's
able to support different outline thicknesses or even no outline,
just for good measure. It's a bunch of work that doesn't change the
way this particular shader works, but will be massively helpful as
a reference for reverse-engineering other shaders and applying the
same correction to prevent division by zero.