Softened autonav speed reset and added an option to soften it further.
Previously, autonav speed would reset if any enemy (whether active
or passive) was within firing range, or if you were in firing range.
It was designed this way to avoid getting the player unexpectedly
destroyed, but it was admittedly extreme, so I've made two changes.

Firstly, enemy firing range is now never considered unless that
enemy is actively hostile and can see the player. Previously, this
condition caused the autonav speed to be reset even when they were
slightly out of range.

Secondly, I added an option called "ignore passive", which causes
all passive enemies (those who aren't actively hostile) to be
entirely ignored for the purposes of deciding whether to reset
speed, meaning a pirate just happening to fly by right next to you
will not slow time down if the pirate is preoccupied with something
else. This option is enabled by default.