New saves now display system name in addition to planet name.
Not sure why this useful information was excluded from the load
screen before, maybe because of how wasteful it used to be? In any
case, it's now included in new saves. (It can't be inferred from the
planet itself due to unidiffs, so old saves continue to not include
this information until they are resaved.)