Restructured pilot name generators and used them for race + skeleton race event.
This was going to be simply duplicating the race missions for a new
race event, but I realized I'm not entirely satisfied with the way
the current missions are, so I need to fix those up first. For now,
this commit contains a migration of those pilot name functions to
a unified module called "pilotname", and migrates everything that
uses those APIs to the new module. I've also added the use of said
API in to the race missions and did some initial restructuring.
This also includes a skeleton race2 event, but it currently does
nothing.