Completely overhauled the Race 1 mission.
I'm not kidding when I say this took literally all day, but it's
much nicer. Will make the same changes to Race 2, and this new race
mission concept will be the basis for the race event.

Will likely also improve Race 1 further. Due to a lack of time, I
haven't yet put much in the way of flavor text.