Removed hilighted pilot text display.
Now that it's literally the only overlay text that conflicts with other
overlay text, it's really growing cold on me. The fact that the names
overlap means missions and events have to be designed with the assumption
that this text can't be seen anyhow, and it just usually isn't needed.
Any place that the text is needed for something, we should compensate
some other way (e.g. by using system marks).