Renamed the "naev" module to "naik".
Just more fully divorcing Naikari's API from its connection with
Naev. It's not strictly speaking necessary, but since it was easy
to do, I figured I might as well make the change. Chose "naik"
instead of "naikari" because it's the same length as "naev".

While I did this, I deleted the LÖVE stuff. We don't use it and I
don't even know if it works.