Reworked civilian NPCs.
Civilian NPCs are now always nongeneric (no distinction between e.g.
"Civilian" and "Imperial Civilian"), and they always attempt to
select a reward to give to the player, so rewards should be much
more common now. Also got rid of the tutorial NPC texts; NPC text is
now purely a vehicle for lore bits. Number has also been reduced to
between 1 and 4 (rather than 4–16), which is slightly smaller than
the 1–5 range that was previously used before it was increased.