Brought race2 up to speed with race1 and lowered race1 reward.
The race2 update hasn't been tested yet, so it might be broken, thô
I hope not. This implementation always gives the player a reward even
if they lose, but a small reward, and has not just a first place prize
but also second and third place prizes.

I still want to expand this, but I don't wanna burn myself out, so
starting with this much. In particular, I wanna add like a form of
audience reaction to make it feel more lively. A bit more ambitiously,
I kinda wanna see about planting a seed for allowing betting on race
events by having race2 include indication of the audience placing
bets.