Adjusted colors to make them easier on the eyes.
Debilitated over this for awhile. Ended up deciding that on the overlay,
planets and jumps are rendered with font colors rather than standard
colors. (This effectively means that the standard colors are now only
used for pilots, where immediately recognizing that a pilot is hostile
to you is more important.)

Also adjusted some of the basic color palette so that all of them
can properly serve as hilight colors. I made a point to make it so
that all of them are clearly distinct from white even with full
colorblindness.