Made gl_Matrix4 functions modify in-place.
This was an attempt to boost performance. It didn't actually make any
difference that I was able to notice, but I figured I'll keep the
change anyhow, because it doesn't seem to hurt performance either.
It was only after I did the work that I realized this was a fool's
errand: these graphics-related things are always going to happen at
the same frequency regardless of time compression, so they're just
not ever going to be a significant factor.