Player is aggressive. Civilians, miners, and traders are non-protectors.
Making the player aggressive is preparation for a possible idea I'd
like to explore in the future: allowing the player to transmit
distress signals. If that were to be done, the player would need to
be marked as aggressive to make sure it isn't unreasonably exploitable.

As for civilians, miners, and traders, marking them as non-protectors
is for the benefit of their fighter bay escorts (which, unlike them,
are aggressive).