Quicksilver, Koäla, and Mule now all have three weapon slots.
Quicksilver and Mule used to have three, but Quicksilver's was dropped
I think so it wouldn't outshine the Koäla, and Mule's was dropped I
think because the Rhino's other benefits seemed real minor. The
Rhino concern is certainly obsolete now (I believe that came from a
time when big weapons weren't really all that much better than medium),
and as for worry about the Koäla, giving it a third weapon slot as
well should address that.

Also updated the Koäla's description.