Fixed options menu calling for restart with zoom changed, simplified zoom.
Near zoom followed this weird formula based on speed or something
which was unintuitive and unnecessary, really only possibly serving
to cause motion sickness during time compression.

Calling for a restart was only necessary because stars and dust needed
to be regenerated, so I made it so it just does this when zoom_far
(the only one which affects these) is changed, and when saving the
options.

Also fixed an issue where dust wasn't recalculated with size changes.