Removed some broken code in cargo_common which would break cargo missions.
This "difference()" function, if it worked properly, would actually
prevent cargo missions from ever delivering to places that have the
commodity service. I believe it doesn't actually do this, thô,
perhaps because of the fact that it tries to use userdata to index a
table (and perhaps this leads to the code not doing what it was
written to do). It's not behavior we want (a leftover from when
different planets had different commodities available), so I'm
removing it to make sure it doesn't end up breaking cargo missions
if the function suddenly were to start working as its author
intended.