Fixed some issues in teenager mission and a confusing idle hook detail.
The way it was coded before, the "idle" hook only actually did anything
if the controlled pilot just *became* idle after having a task. This
caused a bug when I, not realizing this behavior, removed a function
call from the teenager mission that I thought was unnecessary. I see
no legitimate reason for the "idle" hook to behave like this, so I've
changed it so that the "idle" hook gets called if the controlled pilot
is idle, regardless of when it became idle. (This means it'll get
called continuously until the pilot is given a task, or removed from
manual control.) I don't think this will cause any negative consequences,
thô I'll be keeping an eye out to make sure.

I also noticed that movement without compensating for velocity would
lead the Credence to fly around in circles sometimes, so I removed
the disabling of velocity compensation.