Fixed Teenager's use of markers and patched some issues with the music.
The system for switching between ambient and combat music had some
holes in it as pertained to hostiles getting disabled and undisabled.
I've patched those issues; this prevents combat music from playing
when it shouldn't. Also fixed the teenager mission's use of markers
so that it gives one at the start and then moves it, rather than only
giving one after boarding.