Fixed a couple things.
Forgot to commit the first change, so I guess this commit has two fixes.

Firstly, I fixed the Credence in the teenager mission being much too
fast, since cores were unequipped, and made its armor weaker by giving
it Unicorp D-2 plating instead of X-2.

Secondly, I fixed a mistake where the hp03 comm event didn't check
to make sure that the pilot it picked was a natural pilot, which could
cause problems (most notably leading to escorts, which you can't hail,
getting chosen), and while I was at it, I got rid of the unnecessary
system claim (unnecessary as long as it makes sure to only pick a natural
pilot). Also made it return the appropriated ship as a natural pilot
after it's done with it, freeing it up if another event wants to use
it or whatever.