Fixed Turret Conversion Module adding CPU and added limit for it.
We accidentally put the CPU cost in as a positive number, which made
the Turret Conversion Module add CPU instead of costing it. Also added
a limit under the "Weapon Conversion" category. It doesn't really
matter much now, but this prevents the player from doing something
which only has a downside and may be necessary to prevent issues if
other kinds of weapon conversions are added later on.