Made outfit CPU usage an abs value and changed display to "CPU Usage".
This should prevent the error we made with the Turret Conversion Module
from ever happening again. I also changed it so that the CPU usage
display for outfits says "CPU Usage", rahter than a negative value
and "CPU". This will be more intuitive, I think, and fits better
with the mass display (which is similar).