Patched some more potential issues in the music selection code.
I came across another instance of combat music playing when it
shouldn't. I couldn't run any tests to verify, but my best guess for
the cause is maybe an enemy became hostile just as it got disabled,
since I noticed a case where that could cause combat music to play
inappropriately. That's patched, we'll see if that fixes the issue
I ran into.