I think I patched all the music issues, finally.
I ended up doing a massive overhaul of the music system.

The main issue we seem to have been facing before is simply a timing
issue, because the music script wasn't saving the last music unless
it actually played. What I ended up doing is just removing this
distinction; last music is now *always* saved, regardless of whether
it played or not, since the important thing it stores is what music
*should* be playing next time new music is needed.

Also fixed an issue where the music wouldn't change properly when
taking off (since player.isLanded() couldn't actually be reliably
used); this was fixed by passing a variable when the player is
landed rather than relying on the Lua API.

Other than these fixes, I also got rid of takeoff music from the API
(it hasn't actually been in use for a long time), and I changed a
certain code path that caused pilots to get disabled briefly when
they died (which is apparently a holdover from combat rating, a
mechanic that Naikari has never had), since this was seemingly a
key contributing factor to the bugs in the music playback and is
not needed. (It also had the potential to needlessly confuse missions
and events that use disable hooks.)