Slightly adjusted the pilot_boardStealCargo code.
This was because I wrongly thought there was a bug in the code, and
I was simplifying it in one area in the process of figuring it out.
There wasn't actually a bug, but keeping the code edits since they
make it cleaner. This also includes a correction to a comment.