Tied the Nexus campaign to the Baron campaign and adjusted Shark's story.
This makes it so that you're chosen for the Shark demonstration
specifically because you've helped the Baron before and can hopefully
convince him to hear Nexus out on their sales pitch.

This also removes the third Nexus mission (as I don't think it really
fits all that well), lowers the chance of Shark appearing to 20%
rather than 100%, and halves the reward for both remaining Nexus
missions.