Added use of a portrait for Baron Sauterfeldt and adjusted Baron mission.
I needed to adjust the Baron mission to make the description of his
appearance consistent with the new portrait we made for him, so I
also made some other updates, like telling the player about the reward
up-front.

This also adds use of the baron portrait to the Shark mission.