Adjusted things in aiL_minbrakedist to make it more readable.
This doesn't change its function at all, but hopefully should lay
things out in a way that can make it easier to verify what the
correct thing is to do in order to solve #18.

It's that "speed" variable I'm concerned about. It seems like it
shouldn't be there, but presumably it was put there for a reason,
and the fact that a case is leading to apparently zero (or even
negative) speed is itself concerning.