Reverted the change that caused speed to be set to speed_limit.
This particular change caused a regression where those low speeds
accidentally got doubled. The change was an early attempt to fix
issue #18 and should no longer be needed, since the hyperspacing
code is now able to handle the case which caused the problem. That
said, we probably need to add similar handling code to the landing
AI code, which is likely affected in the same way, so in the short-term
I expect this change to cause a worse regression where ultra-slowed
AI pilots will now be unable to land, until we add that updated logic.