Reworked all the AI land code so that it can use planet data, not just pos.
This massive rework (which I hope didn't break anything) means that
the code can now check the planet's radius and use that to make
decisions. It now does so both for deciding whether to afterburn and
for deciding when to start the actual landing procedure, as in the
hyperspacing code.