Removed the "Buy Local Map" button.
It's a bit of a risk, but I think this is the right call in the
long-run, and that we're ready to subject the player to sometimes
having to search for their destination. Local maps are now only
available at planets with outfitters.

The purpose of this change is to make exploration a part of the
game. With standard maps literally everywhere, that's not really
possible anywhere landing is possible. It also increases the
importance of the jump point reveals from NPCs at the bar, thô this
is a minor consideration.