Improved local map description and removed some unused data.
I don't know when local maps had a specified limit to what they
could detect, but that's long-gone, and it isn't even compatible with
how things are done now. So I've removed that obsolete data.

Also touched up the description of the local map to further clarify
exactly what it tells you if you buy it.