gl_program_vert_frag: abort on failure rather than just warn.
This, too, comes from ids1024's GLES conversion, thô he added an
abort() call on top of the WARN() call rather than just changing it
to ERR() for some reason. I'm guessing this was because of getting
bombarded with repeat errors when this problem showed up. I think
keeping that as a fatal error is the right call, since the display
will get horribly messed up if the shaders don't work properly
anyhow.