Removed the Hitman missions.
I was just testing these and find them to be rather tedious and not
worth the trouble. We're also already repurposing the portrait these
missions use for a much better set of missions in Hakoi Pirates, so
these would have needed a major or removal anyway, and I think removal
is the better option.