Improved the autonav bugfix.
That last one was a bit sloppy since if you continually double-clicked
on your jump point, it'd obnoxiously restart the whole thing over
and over again. I changed some things up to avoid that, and this also
fixes an issue where pressing J to cycle jump points was needlessly
finnicky.