Improved the logic which handles low tc_range in sound_setSpeed.
The previous method was a kind of nuclear approach and used values
incorrectly. This corrects the method and makes it a bit more
graceful: instead of disabling volume lowering, it just artificially
sets the range to a minimum value, which gradually reduces the
maximum volume lowering as the actual range gets further from the
required range.