Added a floor to how low kills will take your reputation.
This is kind of like a softer and simpler version of what we tried
to do previously. It should mean that once you get to -20% rep,
self-defense won't cause further reputation loss. However, this
doesn't apply to distressing, so if you are the aggressor, reputation
loss will still happen, albeit a reduced amount.