Touched up the way pilot-related autonav handles target issues.
I noticed a bug where attempting to clear the target while trying
to board would just cause it to retarget it and keep trying. This
was because it was set to the player as the target, which in the
board code was interpreted as meaning board nearest. I changed it
so that instead, during autonav, lacking a target just aborts autonav.
I also made it abort autonav when the target can no longer be boarded,
which also prevents some weirdness where if you switch target while
in autonav board approach, it would then attempt to board and fail
when you arrived at the ship's location.