Increased the penalty for going over mass limit.
The math in the previous penalty was mostly fine, but rather light
all things considered. I decided to strengthen the penalty by
cubing the results. Based on my tests, this seems to always make
the engine that is supposed to make sense actually make sense, and
it makes it impossible for large ships to use small or medium engines
without the assistence of thrust bracers.