Changed the way "dirty" info is used to prevent problems with releases.
Appending ".dirty" onto a release number like "0.11.0" is invalid in
semver, yet that's exactly what gen_version.sh would do if any change
was made to the directory in a Git context. Solved this by making it
include a ".clean" field normally, replace the ".clean" field with
".dirty" if "-dirty" was printed by git describe, and then if "-dirty"
still exists, simply erasing it (meaning release builds will never be
shown as "dirty").

I found out the hard way that this matters for Flathub; the presence
of a patch there caused the game to not run.