Changed a questionable stats recalculation method.
There was some code that, for whatever reason, in stats recalculation
(after e.g. new equipment or activated outfits) caused shield, armor,
and energy values to be modified to be the same "proportion" of the
maximum value. This makes no sense, really. In most cases it simply
meant nothing, but if e.g. an activated outfit had existed that changed
max shield, this quirk in the code would have led to *current* shield
getting changed at the same time, which would almost certainly not be
what was actually intended. I changed it to leave the armor, shield,
and energy values alone, except to cap them to the new max, which
makes a lot more sense.