Changed the way Scout class ships are equipped to prevent fallbacks.
The way it was configured before had a pretty high likelihood of
generating an unspaceworthy result. I've changed it to a different
set which is more in-line with what I actually tend to use myself
when equipping a Scout. This method uses a Unicorp PT-80 Core System
with ultralight engines and hulls, which enables turreted weaponry
and a medium-sized utility while keeping extremely light for fast
movement.