Stat text: changed colors to blue and yellow rather than green and red.
Green and red aren't colorblind accessible, and while fallback symbols
exist, it's better to make the color difference visible to a wider
range of people. This also changes the symbol for stat reduction from
"!!" to "*", and defines red and the "!!" symbol for "dangerous" stats
(that is to say, where a stat is outside the acceptable threshold).