Added some special code that passes along voices from ended beams.
This is to address the problem caused by the sound cancellation code
which would lead to beams going silent just because *other* beams
turned off. This could be seen easily when you had multiple beam types
equipped, some of which had a minimum duration.

The code seems to work perfectly as far as I can tell.