Corrected the vast majority of variables that interacted with SDL_GetTick().
SDL_GetTick() outputs a Uint32 variable, but the vast majority of
variables it and numbers derived from it were assigned to were of
the unsigned int type. This is wrong and in theory could cause needless
problems; the unsigned int type is only guaranteed by the C standard
to be at least 16-bit, and storing time in milliseconds in a 16-bit
variable would cause the time to wraparound roughly every minute
(as opposed to roughly every 49 days on a 32-bit integer). Encountering
a situation where a compiler would actually cause the unsigned int type
to be 16-bit is highly improbable, but not impossible, and I don't
like depending on "probably" for the fundamentals of the code to work.

(This change was made because I noticed Naev switching to the newer
SDL_GetTick64, but we won't be doing this right now because it breaks
compatibility with SDL <2.0.18, which we still want to support. The
possibility of the game glitching slightly after running for 49 days
straight – this would likely manifest as a sudden jump of max delta
timing as the game thinks a lot of time has passed for one frame –
doesn't concern us.)