Lots of structural code changes plus some smaller real changes.
I ended up getting started on a path of wanting to fix the absurdity
of credits-related stuff being in commodity.c/h despite having nothing
to do with commodities, which was a leftover from a previous
economy.c/h which included commodity stuff as well as credits handling.
This took way longer than I thought it would, partly because I got
sidetracked several times.

I also made a couple of small real changes: some text changes, and
a bit more significantly, a change to the color used to represent
that a license or price prevents the player from buying something.
It's now signified not by red text, but by yellow text plus an asterisk.