Changed TC display to be relative instead of absolute.
This means e.g. that what was previously displayed as "TC 200%" is
now displayed as "TC +100%". I think this is a more natural and
intuitive display of what TC does.

This change would have made the previous default TC limit rather
weird in the Options menu, so I changed it from 200 to 201 (which
will be rendered in the Options menu as "+20000%").