Changed colors used for weapon sets in Ship Computer.
I changed it so that black is used for irrelevant outfits, a very
dark red is used for relevant but unused outfits, and two different
shades of blue are used for used outfits. I also re-added the part
of the legend that indicates the color used for unused (but available)
outfits.

With these new colors, the difference between "relevant" and
"irrelevant" slots is very subtle (and almost invisible with
monochromacy), but more noticeable for most types of
color vision, particularly typical color vision and tritanopia.
The colors indicating slots used by the current weapon sets, meanwhile,
are much easier to see with all types of color vision, including
monochromacy; the brightness and hue is very noticeably different.