Completely overhauled slot specialness indicator colors.
I wanted to come up with a better method for colorblind-accessibility,
and I seem to have found one: I'm making use of the whole "light color"
(sheen) system in the toolkit. Using this, exclusive and typed slots
are both a lighter blue than required slots, but have a cyan sheen on
either the top or the bottom. This looks very different depending on
your color vision type, but the result is that they are all very
distinct for all color vision types.

This forced another change: in the shipyard, it now shows similar
bordering, rather than the previous color triangles. This actually
improves accessibility there for other reasons (the triangles were
always hard to see), so that's a win-win in my book.